Move checkpoint code to a package for reuse
