+++ checkpoint/checkpoint_darwin.go
@@ -1,4 +1,4 @@
-package main
+package checkpoint
 
 import (
 	"syscall"
