+++ checkpoint/checkpoint_linux.go
@@ -1,4 +1,4 @@
-package main
+package checkpoint
 
 import (
 	"syscall"
